requirements.txt
# Core dependencies
openai>=1.0.0,<2.0.0
langchain>=0.1.0,<0.2.0
langchain-openai>=0.0.1,<0.1.0

# HTTP and networking
httpx>=0.28.0
httpcore>=1.0.0
anyio>=4.0.0

# Data processing
pandas>=2.0.0
pydantic>=2.0.0

# Utilities
python-dotenv>=1.0.0
tenacity>=8.1.0,<9.0.0

# Visualization
plotly>=5.18.0
kaleido>=0.2.1  # Required for static image export

# Development and testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Logging
structlog>=23.0.0

# Type hints
typing-extensions>=4.0.0
